py-symbol: Allow gdb_lookup_symbol to accept block=None
gdbpy_lookup_symbol refuses to perform a lookup without a valid block.  If
no block is passed, we look up the block using the selected frame.  If there
is no selected frame, we return an error.  Since a frame isn't required
to lookup objects like static functions and variables, we should be
able to complete the lookup without a block.  This patch allows us to
accept block=None to proceed with the lookup without a block.

This allows us to lookup static symbols directly from python rather
than using gdb.eval_and_parse.