gdb-fortran-frame-string.patch
;; Display Fortran strings in backtraces.
;;=fedoratest

http://sourceware.org/ml/gdb-patches/2014-07/msg00709.html

Hi,

for Fortran it fixes displaying normal strings also in frames/backtraces:

(gdb) frame
->

The patch is simple and I do not see why it should not be this way.

For C/C++ TYPE_CODE_STRING is not used.  I am not aware of Pascal but that
language is currently not really much supported in GDB anyway.

This was a part of my archer/jankratochvil/vla branch but it is not a part of
the Intel VLA patchset as it in fact is completely unrelated to "VLA".

No regressions on {x86_64,x86_64-m32,i686}-fedora22pre-linux-gnu.

Thanks,
Jan