gdb: properly zero out memory used for fields when allocated
The memory allocated for fields in process_enumeration_scope is extended
with a xrealloc call in chunks but the memory is never fully initialized
and the fields are only partially written.  We then copy only the written
fields into the type, where later we assume everything was initialized
properly.  This mostly works until we want to do deep type comparison
and the type comparison fails.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>