gdb: exclude PROP_UNDEFINED ranges from bound_flag_evaluated check
The bound_flag_evaluated flag will be set when a dynamic type is evaluated
as a static type.  If the high range is PROP_UNDEFINED, nothing will
be changed when it's evaluated, but the equality check will still fail.

This makes it impossible to compare types in Python code if the type
contains a zero-length array.

Acked-by: Jeff Mahoney <jeffm@suse.com>