target: add delete_target
The forthcoming python target will allow the user to register a target
implementation in Python.  Since these are not static targets, it would
make sense to allow the target to be unregistered from gdb.

This commit adds a new delete_target() call that removes the target
from the target factories and removes the target subcommand to access it.
The target itself will be cleaned up when the last reference to it is dropped.