py-register: Add interface to registers
This patch adds new a gdb.Register object to access the register cache
associated with a thread.  Registers may not be created independently but
may be accessed via the gdb.InferiorThread.registers dictionary.  Registers
are invalidated when the thread exits much like symbols are invalidated
when an objfile is unloaded.