py-lktarget: add small linux kernel target
This target implements some basic semantic debugging that should apply
to most kernels.  For now it just relocates the kernel and populates
the thread list for handling by a derived target written in Python.
Things that are slow to do in Python (like iteration of big lists, etc)
should be implemented here in helpers.